+++ src/game/world/renderer/SinusGraph.cpp
@@ -9,21 +9,21 @@ SinusGraph::SinusGraph() {}
 SinusGraph::~SinusGraph() {}
 
 // Gibt die Höhe von der X Position an, ermittelt aus der Funktion f(x)
-float SinusGraph::heightFunction(float valueX) {
+double SinusGraph::heightFunction(float valueX) {
     return 0.025 * sin(valueX) * valueX;
 }
 
 // Gibt die Steigung von der X Position an, ermittelt aus der Funktion f'(x)
-float SinusGraph::heightFunctionDerivation(float valueX) {
+double SinusGraph::heightFunctionDerivation(float valueX) {
     return 0.025 * cos(valueX) * valueX + 0.025 * sin(valueX);
 }
 
 // Gibt die Höhe von der Z Position an, ermittelt aus der Funktion g(x)
-float SinusGraph::depthFunction(float valueZ) {
+double SinusGraph::depthFunction(float valueZ) {
     return -0.5 * valueZ * valueZ;
 }
 
 // Gibt die Steigung von der Z Position an, ermittelt aus der Funktion g'(x)
-float SinusGraph::depthFunctionDerivation(float valueZ) {
+double SinusGraph::depthFunctionDerivation(float valueZ) {
     return -1 * valueZ;
 }
